feat: se agrego el contexto global para el carrito de compras local
