+++ db.json
@@ -2,7 +2,7 @@
   "products": [
     {
       "id": "1",
-      "name": "Producto AAA",
+      "name": "Producto AAAA",
       "description": "Juguete para varones",
       "price": 60,
       "stock": 100,
